Refactor: Split subquery_rewrite.rs into modular components (#2185)
Split the 1066-line subquery_rewrite.rs file into focused modules:

**New module structure:**
- `detection.rs` (~145 lines): IN subquery detection logic
- `correlation.rs` (~210 lines): Subquery correlation analysis
- `transformations.rs` (~100 lines): IN→EXISTS and DISTINCT transformations
- `expression.rs` (~270 lines): Recursive expression rewriting
- `mod.rs` (~340 lines): Public API and tests

**Benefits:**
- Single Responsibility: Each module has one clear purpose
- Easier Testing: Can test correlation independently from transformations
- Better Documentation: Focused modules with clearer API boundaries
- Reduced Cognitive Load: Each file under 350 lines (from 1066)
- Improved Maintainability: Changes isolated to relevant modules

**Testing:**
- All 11 subquery_rewrite tests pass
- All 1032 vibesql-executor tests pass
- No change in functionality or public API

Closes #2185
